+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## HEAD
 
+### Changed
+- [modules] Simplified code related to caching of `data.SITE.matter` and `data.SITE.images`.
+
 ### Fixed
 - [modules] Replaced ES6 imports and exports in some module files with Node `require` and `module.exports`.